rename(route-folder): 공지사항 컨벤션 반영
