+++ src/app/(route)/notice/[id]/_components/NoticeDetailView/_internal/NoticeDetailSkeleton/NoticeDetailSkeleton.tsx
