+++ src/app/(route)/notice/[id]/_components/NoticeDetailView/_internal/index.ts
@@ -1,3 +1,3 @@
-export { default as NoticeDetailContent } from "./NoticeDetailContent";
-export { default as NoticeCommentForm } from "./NoticeCommentForm";
-export { default as NoticeDetailSkeleton } from "./NoticeDetailSkeleton";
+export { default as NoticeDetailContent } from "./NoticeDetailContent/NoticeDetailContent";
+export { default as NoticeCommentForm } from "./NoticeCommentForm/NoticeCommentForm";
+export { default as NoticeDetailSkeleton } from "./NoticeDetailSkeleton/NoticeDetailSkeleton";
