--- src/app/(route)/notice/_constant/commentListObject.ts
@@ -1,28 +1,0 @@
-export const commentListObject = [
-  {
-    id: 1,
-    author: "작성자1",
-    date: "2025.05.06",
-    content:
-      "서울시 노원구 00동 건물 화장실에서 핸드폰을 분실했어요. 혹시 발견하신 분이 계시면 채팅 부탁드려요.",
-  },
-  {
-    id: 2,
-    author: "작성자2",
-    date: "2025.05.06",
-    content: "저도 비슷한 경험이 있었는데 근처 경비실에 맡겨져 있을 수도 있어요.",
-  },
-  {
-    id: 3,
-    author: "작성자3",
-    date: "2025.05.07",
-    content: "핸드폰 켜져 있으면 '나의 아이폰 찾기' 같은 기능 활용해보세요!",
-  },
-  {
-    id: 4,
-    author: "작성자4",
-    date: "2025.05.07",
-    replyTo: "작성자1",
-    content: "저는 근처 화장실에서 비슷한 핸드폰을 본 것 같아요. 확인해보시는 게 좋겠습니다.",
-  },
-];
